requirement.txt
tensorflow==2.10.0
scikit-learn==1.3.0
numpy==1.24.3
matplotlib==3.7.2
pandas==1.5.3
streamlit==1.20.0
flask==2.3.3
Pillow==10.0.0
